+++ nirum/rpc.py
@@ -320,12 +320,8 @@ class Client:
         self.opener = opener
 
     def ping(self):
-        req = urllib.request.Request(
-            urllib.parse.urljoin(self.url, './ping/'),
-            headers={'Content-Type': 'application/json;charset=utf-8',
-                     'Accepts': 'application/json'}
-        )
-        return self.make_request(req)
+        r = self.do_request(urllib.parse.urljoin(self.url, './ping/'), {})
+        return json.loads(r) == 'Ok'
 
     def remote_call(self, method_name, payload={}):
         qs = urllib.parse.urlencode({'method': method_name})